requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.2.0
appnope==0.1.3
archspec @ file:///croot/archspec_1697725767277/work
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==6.0.0
astropy-iers-data==0.2023.12.18.0.30.18
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
boltons @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f63n9uulmp/croot/boltons_1677628710094/work
Brotli @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_38mvgltu8c/croots/recipe/brotli-split_1659616064542/work
cachetools==5.5.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_35eq66o3mo/croot/certifi_1700501684871/work/certifi
cffi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b4nang6w_y/croot/cffi_1700254307954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comm==0.2.0
conda @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_f7x3zuzbl2/croot/conda_1699303137115/work
conda-libmamba-solver @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_c2fmjn55mo/croot/conda-libmamba-solver_1700150913462/work/src
conda-package-handling @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_fc4cx8vjhj/croot/conda-package-handling_1690999937094/work
conda_package_streaming @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_aecpaup22q/croot/conda-package-streaming_1690987978274/work
contourpy==1.1.0
cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_905z2r5rpq/croot/cryptography_1694211573866/work
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docutils==0.21.2
dqsegdb2==1.2.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_20tv0lrp8x/croot/filelock_1672387134240/work
fonttools==4.41.0
fqdn==1.5.1
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/croot-muov7gw1/gmpy2_1645462616268/work
google-ai-generativelanguage==0.4.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.62.3
gwdatafind==1.2.0
gwosc==0.7.1
gwpy==3.0.7
h5py==3.10.0
html5lib==1.1
httplib2==0.22.0
idna @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_771olrhiqw/croot/idna_1666125579282/work
igwn-auth-utils==1.1.0
importlib-metadata==7.0.0
importlib-resources==6.0.0
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
Jinja2 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_9fjgzv9ant/croot/jinja2_1666908141308/work
json5==0.9.14
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==25.3.0
kiwisolver==1.4.4
libmambapy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_59l2npsw_8/croot/mamba-split_1698782625405/work/libmambapy
ligo-segments==1.4.0
ligotimegps==2.0.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_12c133f5-0720-4727-9c18-599a3af825723lzwham3/croots/recipe/markupsafe_1654597866058/work
matplot==0.1.9
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
mpmath @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_17iu6a8a3m/croot/mpmath_1690848269369/work
multitasking==0.0.11
narwhals==1.18.4
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b9af3smw_7/croot/networkx_1690562010704/work
nh3==0.2.18
notebook==7.0.6
notebook_shim==0.2.3
numpy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_362zs5g963/croot/numpy_and_numpy_base_1695830450707/work/dist/numpy-1.26.0-cp39-cp39-macosx_11_0_arm64.whl#sha256=e37b41d15b4b8b2adde02a2f2a3e74daaccdfa5962f4470c1c266acd01ffd2a0
overrides==7.4.0
packaging @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6dm6d4jd_t/croot/packaging_1693575176524/work
pandas==2.0.3
pandoc==2.3
pandocfilters==1.5.0
parso==0.8.3
peewee==3.17.6
pexpect==4.9.0
Pillow @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_ccx0_h9rkg/croot/pillow_1696580032390/work
pkginfo==1.10.0
platformdirs==4.1.0
pluggy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_croot-w6jyveby/pluggy_1648109277227/work
plumbum==1.8.2
ply==3.11
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==4.25.5
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycosat @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3eg8vdcs6z/croot/pycosat_1696536519213/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pyerfa==2.0.1.1
Pygments==2.17.2
PyJWT==2.8.0
pyloco==0.0.139
pyOpenSSL @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b8whqav6qm/croot/pyopenssl_1690223428943/work
pyparsing==3.0.9
PySocks @ file:///Users/ktietz/Code/oss/ci_pkgs/pysocks_1626781349491/work
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
readme_renderer==44.0
referencing==0.32.0
requests @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_54zi68h2nb/croot/requests_1690400233316/work
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.15.2
rsa==4.9
ruamel.yaml @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_aeh5mqcw49/croot/ruamel.yaml_1666304555976/work
ruamel.yaml.clib @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f64xdg2rww/croot/ruamel.yaml.clib_1666302244208/work
safe-netrc==1.0.1
scipy==1.11.4
scitokens==1.8.1
seaborn==0.12.2
Send2Trash==1.8.2
SimpleWebSocketServer==0.1.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.41.1
sympy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_21kpmugwgm/croot/sympy_1668202392692/work
tenacity==8.5.0
terminado==0.18.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torchvision==0.15.0
tornado==6.4
tqdm @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ac7zic_tin/croot/tqdm_1679561870178/work
traitlets==5.14.0
twine==5.1.1
types-python-dateutil==2.8.19.14
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_068obtb882/croot/urllib3_1698257558009/work
ushlex==0.99.1
validators==0.34.0
watchdog==6.0.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
yfinance==0.2.43
zipp==3.16.2
zstandard @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_41b3vxtask/croot/zstandard_1677013668452/work
